Move login page to root pages
