+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Login.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model ConcernsCaseWork.Pages.Account.LoginModel
+@model ConcernsCaseWork.Pages.LoginModel
 
 @{
 	ViewBag.Title = "Login";
@@ -14,7 +14,6 @@
 
 <div class="govuk-grid-row">
 	<div class="govuk-grid-column-one-half">
-		<h2 class="govuk-heading-l">Sign in</h2>
 		
 		@if (!string.IsNullOrEmpty((string) TempData.Peek("Error.Message")))
 		{
